Conflicts:
	src/open/core/cameracalibration/calibrationCamera.h
	src/open/core/cameracalibration/calibrationHelpers.cpp
	src/open/core/cameracalibration/calibrationHelpers.h
	src/open/core/cameracalibration/calibrationPhotostation.h
	src/open/core/geometry/conic.cpp
	src/restricted/applications/indoors/indoorLocalizerRayCostFunction.h
	src/restricted/applications/indoors/main.cpp

+++ core/cameracalibration/calibrationScene.cpp
@@ -2,7 +2,6 @@
 
 CalibrationScene::CalibrationScene()
 {
-
 }
 
 void CalibrationScene::projectForward(CalibrationFeaturePoint::PointType mask, bool round)
@@ -21,7 +20,7 @@ void CalibrationScene::projectForward(CalibrationFeaturePoint::PointType mask, b
             for (size_t camId = 0; camId < station.cameras.size(); camId++)
             {
                 CameraModel *camera = &(station.cameras[camId]);
-                CameraModel worldCam = station.getWorldCamera(camId);
+                CameraModel worldCam = station.getWorldCamera((int)camId);
 
 
                 Vector2dd projection = worldCam.project(point->position);
@@ -53,4 +52,3 @@ void CalibrationScene::projectForward(CalibrationFeaturePoint::PointType mask, b
         }
     }
 }
-
